feat: support building multiple bundles
The pact-ruby-standalone project recently added the pact-plugin-cli to the bundle

This weighs in at 12Mb and hugely inflates our bundle.

This proposal suggests a couple of different packages

pact - ruby only pact packages
pact-rust - ruby + rust packages
pact-ffi - ruby + pact_ffi lib and pact-ffi gem
pact-cli - ruby + rust + pact_ffi lib and pact-ffi gem

for any of the combinations above, they can have

pact-slim-<xyz>

augmentations, which seek to trim out as much as possible from the ruby runtime.

Note:- the slim versions may not work and need throughly testing, cursory testing has been
  done so far.

The overall aim is provide users with a single bundled distribution called pact-cli

which contains every single pact executable so the users can get all the pact goodness at their
  fingertips.

This will have benefits for packaging with HomeBrew/Scoop/Choco and our own install script.

In the future it would be nice to extend pact-plugin-cli which can discover and install plugins

to discover all of the pact tooling (official and community provided) and provide a mechanism to install

to a shared location such as PACT_HOME_DIR which would default to $HOME/.pact

This pattern is already in use in the pact-plugin-cli

Could we further this, so users only need the standalone in a single location, and the ffi libs in a single location, and our various client libraries to point to the relevant home folder, and not require the packages to contain the host libraries themselves. This may not be practical for some libs which need to build from source, but if the source was already on their machine, via our installer, maybe that makes it less of an issue.

The tree of all tools would currently look like this

```lisp
├── pact
├── pactflow
├── pact-broker
├── pact-message
├── pact-mock-service
├── pact-stub-service
├── pact-provider-verifier
├── pact_mock_server_cli
├── pact-stub-server
├── pact_verifier_cli
└── pact-plugin-cli
```

Users would be able to interact with the FFI by

    ${PATH_TO_RUBY}ruby -rpact/ffi -e "puts PactFfi.pactffi_version"